fix: use vitest test.projects with root coverage config
Replace vitest.workspace.ts with test.projects in vitest.config.ts
so coverage config lives at the root level (where vitest expects it)
while still splitting unit (node) and component (jsdom) environments.
Fixes CI failure where coverage-summary.json was missing.
