+++ package.json
@@ -9,6 +9,8 @@
     "lint": "eslint --cache",
     "type-check": "tsc --noEmit",
     "test": "vitest run",
+    "test:unit": "vitest run --project unit",
+    "test:component": "vitest run --project component",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
     "sync": "tsx scripts/sync-dreps.ts",
